requirements.txt
# Apple Silicon TensorFlow support
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0

# Core data science packages
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
Pillow==10.3.0
tqdm==4.66.1
solara==1.49.0

# For saving/loading training history (optional)
pickle5==0.0.11

# TensorBoard for visualization
tensorboard==2.15.1

# Jupyter support (optional)
notebook==7.2.0
ipykernel==6.29.4